+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 5.4.9 (unreleased)
 ------------------
 
-- Limit event occurrences to 1000.
+- Limit event occurrences to 100.
   [mamico]
 - Customize INameChooser adapter to check also alias ids and disallow to create contents that could override aliases.
   [cekk]